+++ CHANGELOG.md
@@ -9,3 +9,16 @@
  - Launch of nuxt-sections module migrated from vue-sections library
 
  - Library updates needed after the backend fixes done on media #93
+
+
+### v1.0.1 (23/5/2023)
+
+---
+
+ - Bug fixes for the media content linking of configurable sections
+
+ - Support for cookie-universal-nuxt alias to dynamically read the alias used by the host project
+ 
+ - SEO metadata management #101
+
+ - New feature implementation of having default support for dynamic pages